requirements.txt
# IMPORTANT: This project requires PyTorch NIGHTLY BUILD for RTX 5070 support
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu129

# Core PyTorch and Deep Learning (CUDA 12.9 Nightly)
# DO NOT install these from regular PyPI - use nightly index
torch==2.9.0.dev20250831+cu129
torchvision==0.24.0.dev20250831+cu129
torchaudio==2.8.0.dev20250831+cu129
torch-tensorrt==2.9.0.dev20250831+cu129

# Scientific Computing
numpy==2.3.2
scipy==1.16.1
matplotlib==3.10.6

# Machine Learning Framework
transformers==4.56.0
accelerate==1.10.1
peft==0.17.1
einops==0.8.1
safetensors==0.6.2
huggingface-hub==0.34.4
tokenizers==0.22.0

# Experiment Tracking
wandb==0.21.3

# Model Optimization
tensorrt==10.12.0.36
tensorrt-cu12==10.12.0.36
tensorrt-cu12-bindings==10.12.0.36
tensorrt-cu12-libs==10.12.0.36
nvidia-cuda-runtime-cu12==12.9.79
nvidia-modelopt==0.27.0
nvidia-modelopt-core==0.27.0

# Visualization and Plotting
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
kiwisolver==1.4.9
pillow==11.3.0
pyparsing==3.2.3

# Configuration and Data
PyYAML==6.0.2
lxml==6.0.1
python-pptx==1.0.2
xlsxwriter==3.2.5

# Web and API
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# Utilities
tqdm==4.67.1
click==8.2.1
colorama==0.4.6
filelock==3.19.1
fsspec==2025.7.0
packaging==25.0
psutil==7.0.0
six==1.17.0
python-dateutil==2.9.0.post0
torchsummary==1.5.1
dllist==2.0.0

# Development Tools
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2
platformdirs==4.4.0
sentry-sdk==2.35.1
protobuf==6.32.0
regex==2025.8.29
ninja==1.13.0

# Type Checking and Validation
pydantic==2.11.7
pydantic-core==2.33.2
annotated-types==0.7.0
typing-extensions==4.14.1
typing-inspection==0.4.1

# Math and Symbolic Computing
sympy==1.14.0
mpmath==1.3.0
networkx==3.5

# Rich Console Output
rich==14.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2

# Template Engine
Jinja2==3.1.6
MarkupSafe==3.0.2